+++ docs/source/changelog.rst
@@ -6,5 +6,5 @@ Release 0.1.0
 
 `9 September 2017`
 
-* Added basic Vector class
+* Added basic Vector class.
 * Added basic degrees/radians conversion decorator.